requirements/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
